+++ CHANGELOG.md
@@ -15,12 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add types required in `AudioOutputProvider`
 - Add sound disabled to `AudioOutputControl`
 - Add `LocalAudioOutputProvider` to `MeetingView`
+- Add `ContentShareControlProvider` for content share control methods
 
 ### Changed
 - Added delay for more consistent animation snapshotting with playwright.
 - `npm run test` script changed to only include /components and /utils
 - Update prebuild script to trigger git actions on git push event
 - Changed relative imports and `MeetingView` to accomodate imports of connected components from library
+- Updated the `ContentShareControl` with SDK components
+- Changed `ScreenShare` to `ContentShare` in library and demo
 
 ### Removed
 - Removed active state button tests.